test: shard tui gateway s5 clusters
